table backend

+++ app.py
@@ -13,7 +13,7 @@ def calculate():
     try:
         initial_capital = float(data.get('initialCapital', 0))
         final_capital = float(data.get('finalCapital', 0))
-        num_periods = float(data.get('numPeriods', 0))
+        num_periods = int(data.get('numPeriods', 0))
         periodic_contribution = float(data.get('periodicContribution', 0))
         frequency = data.get('frequency')
         if initial_capital < 0 or final_capital < 0 or num_periods <= 0 or periodic_contribution < 0:
